vecindex: prepare to add fixup workers
This is a purely mechanical commit that splits some of the fixup code into
a new fixup_worker.go file. This makes it easier to review fixup worker
support in subsequent commits.

Epic: CRDB-42943

Release note: None
