+++ pkg/sql/vecindex/fixup_worker_test.go
